Fix partial month date ranges and data gaps
Extract get_start_dates_for_date_range into a separate function
and add a test for it to ensure it behaves as expected. Start
and end date parameters should be extended to the start of the
month and start of the following month (if the end date is not
on the start of a month already) respectively.

Also handle month+ data gaps between start and end dates of a dataset,
which will cause ERDDAP to return 404 when that month is requested.